+++ src/app/app.module.ts
@@ -10,6 +10,7 @@ import {ProductModule} from './product/product.module';
 import {OrdersModule} from './orders/orders.module';
 import {AdminModule} from './admin/admin.module';
 import {TimingInterceptor} from 'src/app/core/interceptors';
+import { RootStoreModule } from './core/@ngrx/root-store.module';
 
 @NgModule({
     declarations: [AppComponent],
@@ -21,7 +22,7 @@ import {TimingInterceptor} from 'src/app/core/interceptors';
         OrdersModule,
         LayoutModule,
         AdminModule,
-
+        RootStoreModule,
         AppRoutingModule,
     ],
     providers: [{