+++ src/app/cart/cart-routing.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CartListComponent } from '.';
+import { CartStatePreloadingGuard } from '../core/guards';
 
 const routes: Routes = [
     {
         path: 'cart',
         component: CartListComponent,
+        canActivate: [CartStatePreloadingGuard],
     },
 ];