+++ src/app/core/interceptors/timing.interceptor.ts
@@ -12,7 +12,7 @@ export class TimingInterceptor implements HttpInterceptor {
                 .pipe(
                     map((event: HttpEvent<any>) => {
                         if (event instanceof HttpResponse) {
-                            console.info(`TimingInterceptor: ${req.url} took ${performance.now() - start} ms`);
+                            console.log(`TimingInterceptor: ${req.url} took ${performance.now() - start} ms`);
                             return event;
                         }
                     }));